+++ engine/api/workflow_queue.go
@@ -1058,6 +1058,111 @@ func (api *API) postWorkflowJobTestsResultsHandler() service.Handler {
 	}
 }
 
+func (api *API) workflowRunResultCheckUploadHandler() service.Handler {
+	return func(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+		jobID, err := requestVarInt(r, "permJobID")
+		if err != nil {
+			return err
+		}
+
+		if !isCDN(ctx) {
+			return sdk.WrapError(sdk.ErrForbidden, "only CDN can call this route")
+		}
+
+		wr, err := workflow.LoadRunByJobID(api.mustDBWithCtx(ctx), jobID, workflow.LoadRunOptions{DisableDetailledNodeRun: true})
+		if err != nil {
+			return err
+		}
+		proj, err := project.LoadByID(api.mustDB(), wr.ProjectID)
+		if err != nil {
+			return err
+		}
+
+		_, enabled := featureflipping.IsEnabled(ctx, gorpmapping.Mapper, api.mustDB(), sdk.FeatureCDNArtifact, map[string]string{
+			"project_key": proj.Key,
+		})
+		if !enabled {
+			return sdk.WrapError(sdk.ErrForbidden, "cdn is not enabled for project %s", proj.Key)
+		}
+
+		var apiRef sdk.CDNRunResultAPIRef
+		if err := service.UnmarshalBody(r, &apiRef); err != nil {
+			return sdk.WithStack(err)
+		}
+
+		if apiRef.ToFilename() == "" {
+			return sdk.WrapError(sdk.ErrInvalidData, "unable to read run result api ref")
+		}
+
+		b, err := workflow.CanUploadRunResult(ctx, api.mustDBWithCtx(ctx), api.Cache, *wr, apiRef)
+		if err != nil {
+			return err
+		}
+		if !b {
+			return sdk.WrapError(sdk.ErrInvalidData, "unable to duplicate an artifact")
+		}
+
+		// Save check
+		if err := api.Cache.SetWithTTL(workflow.GetRunResultKey(apiRef.RunID, apiRef.RunResultType, apiRef.ArtifactName), true, 600); err != nil {
+			return sdk.WrapError(err, "unable to cache result artifact check %s ", apiRef.ToFilename())
+		}
+		return nil
+	}
+}
+
+func (api *API) postWorkflowRunResultsHandler() service.Handler {
+	return func(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+		jobID, err := requestVarInt(r, "permJobID")
+		if err != nil {
+			return err
+		}
+
+		if !isCDN(ctx) {
+			return sdk.WrapError(sdk.ErrForbidden, "only CDN can call this route")
+		}
+
+		var runResult sdk.WorkflowRunResult
+		if err := service.UnmarshalBody(r, &runResult); err != nil {
+			return sdk.WithStack(err)
+		}
+
+		wr, err := workflow.LoadRunByJobID(api.mustDBWithCtx(ctx), jobID, workflow.LoadRunOptions{DisableDetailledNodeRun: true})
+		if err != nil {
+			return err
+		}
+		proj, err := project.LoadByID(api.mustDB(), wr.ProjectID)
+		if err != nil {
+			return err
+		}
+
+		_, enabled := featureflipping.IsEnabled(ctx, gorpmapping.Mapper, api.mustDB(), sdk.FeatureCDNArtifact, map[string]string{
+			"project_key": proj.Key,
+		})
+		if !enabled {
+			return sdk.WrapError(sdk.ErrForbidden, "cdn is not enabled for project %s", proj.Key)
+		}
+
+		if wr.ID != runResult.WorkflowRunID {
+			return sdk.WrapError(sdk.ErrInvalidData, "unable to add artifact on this run: %s", runResult.ID)
+		}
+
+		nr, err := workflow.LoadNodeRunByID(api.mustDB(), runResult.WorkflowNodeRunID, workflow.LoadRunOptions{})
+		if err != nil {
+			return err
+		}
+		if nr.WorkflowRunID != wr.ID {
+			return sdk.WrapError(sdk.ErrInvalidData, "invalid node run %d", runResult.WorkflowNodeRunID)
+		}
+		runResult.SubNum = nr.SubNumber
+
+		if err := workflow.AddResult(api.mustDBWithCtx(ctx), api.Cache, &runResult); err != nil {
+			return err
+		}
+
+		return nil
+	}
+}
+
 func (api *API) postWorkflowJobTagsHandler() service.Handler {
 	return func(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
 		if isWorker := isWorker(ctx); !isWorker {
