+++ engine/api/workflow_queue_test.go
@@ -6,6 +6,9 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"github.com/ovh/cds/engine/api/bootstrap"
+	"github.com/ovh/cds/engine/api/database/gorpmapping"
+	"github.com/ovh/cds/engine/featureflipping"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -1602,4 +1605,158 @@ func Test_postWorkflowJobSetVersionHandler(t *testing.T) {
 	rec = httptest.NewRecorder()
 	router.Mux.ServeHTTP(rec, req)
 	require.Equal(t, 403, rec.Code)
+}
+
+func Test_workflowRunResultsAdd(t *testing.T) {
+	featureflipping.Init(gorpmapping.Mapper)
+	api, db, router := newTestAPI(t, bootstrap.InitiliazeDB)
+
+	feat := &sdk.Feature{
+		Name: sdk.FeatureCDNArtifact,
+		Rule: "return true",
+	}
+	require.NoError(t, featureflipping.Insert(gorpmapping.Mapper, db, feat))
+	t.Cleanup(func() {
+		_ = featureflipping.Delete(db, feat.ID)
+	})
+
+	cdnServices, _, jwtCDN := assets.InitCDNService(t, db)
+	t.Cleanup(func() { _ = services.Delete(db, cdnServices) })
+
+	u, _ := assets.InsertAdminUser(t, db)
+	consumer, _ := authentication.LoadConsumerByTypeAndUserID(context.TODO(), db, sdk.ConsumerLocal, u.ID, authentication.LoadConsumerOptions.WithAuthentifiedUser)
+
+	key := sdk.RandomString(10)
+	proj := assets.InsertTestProject(t, db, api.Cache, key, key)
+
+	w := assets.InsertTestWorkflow(t, db, api.Cache, proj, sdk.RandomString(10))
+
+	wrCreate, err := workflow.CreateRun(api.mustDB(), w, sdk.WorkflowRunPostHandlerOption{AuthConsumerID: consumer.ID})
+	assert.NoError(t, err)
+
+	require.NoError(t, api.workflowRunCraft(context.TODO(), wrCreate.ID))
+
+	wrDB, err := workflow.LoadRunByID(db, wrCreate.ID, workflow.LoadRunOptions{})
+	require.NoError(t, err)
+
+	nr := wrDB.WorkflowNodeRuns[w.WorkflowData.Node.ID][0]
+	nr.Status = sdk.StatusBuilding
+	require.NoError(t, workflow.UpdateNodeRun(db, &nr))
+
+	nrj := nr.Stages[0].RunJobs[0]
+	nrj.Status = sdk.StatusBuilding
+	workflow.UpdateNodeJobRun(context.Background(), db, &nrj)
+
+	//Prepare request
+	vars := map[string]string{
+		"permJobID": fmt.Sprintf("%d", nrj.ID),
+	}
+
+	artiData := sdk.WorkflowRunResultArtifact{
+		Size:       1,
+		MD5:        "AA",
+		CDNRefHash: "AA",
+		Name:       "myartifact",
+		Perm:       0777,
+	}
+	bts, err := json.Marshal(artiData)
+	addResultRequest := sdk.WorkflowRunResult{
+		WorkflowRunID:     wrCreate.ID,
+		WorkflowNodeRunID: nr.ID,
+		WorkflowRunJobID:  nrj.ID,
+		Type:              sdk.WorkflowRunResultTypeArtifact,
+		DataRaw:           bts,
+	}
+
+	uri := router.GetRoute("POST", api.postWorkflowRunResultsHandler, vars)
+	test.NotEmpty(t, uri)
+	req := assets.NewJWTAuthentifiedRequest(t, jwtCDN, "POST", uri, addResultRequest)
+
+	//Do the request
+	rec := httptest.NewRecorder()
+	router.Mux.ServeHTTP(rec, req)
+
+	// Can't work because check has not be done
+	assert.Equal(t, 403, rec.Code)
+
+	// add check
+	require.NoError(t, api.Cache.SetWithTTL(workflow.GetRunResultKey(wrCreate.ID, sdk.WorkflowRunResultTypeArtifact, artiData.Name), true, 60))
+
+	//Do the request
+	reqOK := assets.NewJWTAuthentifiedRequest(t, jwtCDN, "POST", uri, addResultRequest)
+	recOK := httptest.NewRecorder()
+	router.Mux.ServeHTTP(recOK, reqOK)
+	assert.Equal(t, 204, recOK.Code)
+
+	b, err := api.Cache.Exist(workflow.GetRunResultKey(wrCreate.ID, sdk.WorkflowRunResultTypeArtifact, artiData.Name))
+	require.NoError(t, err)
+	require.False(t, b)
+
+}
+
+func Test_workflowRunResultCheckUpload(t *testing.T) {
+	featureflipping.Init(gorpmapping.Mapper)
+	api, db, router := newTestAPI(t, bootstrap.InitiliazeDB)
+
+	u, _ := assets.InsertAdminUser(t, db)
+	consumer, _ := authentication.LoadConsumerByTypeAndUserID(context.TODO(), db, sdk.ConsumerLocal, u.ID, authentication.LoadConsumerOptions.WithAuthentifiedUser)
+
+	key := sdk.RandomString(10)
+	proj := assets.InsertTestProject(t, db, api.Cache, key, key)
+
+	feat := &sdk.Feature{
+		Name: sdk.FeatureCDNArtifact,
+		Rule: "return true",
+	}
+	require.NoError(t, featureflipping.Insert(gorpmapping.Mapper, db, feat))
+	t.Cleanup(func() {
+		_ = featureflipping.Delete(db, feat.ID)
+	})
+
+	w := assets.InsertTestWorkflow(t, db, api.Cache, proj, sdk.RandomString(10))
+
+	wrCreate, err := workflow.CreateRun(api.mustDB(), w, sdk.WorkflowRunPostHandlerOption{AuthConsumerID: consumer.ID})
+	assert.NoError(t, err)
+
+	require.NoError(t, api.workflowRunCraft(context.TODO(), wrCreate.ID))
+
+	wrDB, err := workflow.LoadRunByID(db, wrCreate.ID, workflow.LoadRunOptions{})
+	require.NoError(t, err)
+
+	nr := wrDB.WorkflowNodeRuns[w.WorkflowData.Node.ID][0]
+	nr.Status = sdk.StatusBuilding
+	require.NoError(t, workflow.UpdateNodeRun(db, &nr))
+
+	nrj := nr.Stages[0].RunJobs[0]
+	nrj.Status = sdk.StatusBuilding
+	workflow.UpdateNodeJobRun(context.Background(), db, &nrj)
+
+	cdnServices, _, jwtCDN := assets.InitCDNService(t, db)
+	t.Cleanup(func() { _ = services.Delete(db, cdnServices) })
+
+	//Prepare request
+	vars := map[string]string{
+		"key":              proj.Key,
+		"permWorkflowName": w.Name,
+		"number":           fmt.Sprintf("%d", wrCreate.Number),
+		"nodeID":           fmt.Sprintf("%d", wrCreate.Workflow.WorkflowData.Node.ID),
+	}
+	checkRequest := sdk.CDNRunResultAPIRef{
+		ArtifactName: "myArtifact",
+		RunID:        wrCreate.ID,
+		RunNodeID:    nr.ID,
+		RunJobID:     nrj.ID,
+		WorkflowID:   w.ID,
+		WorkflowName: w.Name,
+		ProjectKey:   key,
+	}
+
+	uri := router.GetRoute("POST", api.workflowRunResultCheckUploadHandler, vars)
+	test.NotEmpty(t, uri)
+	req := assets.NewJWTAuthentifiedRequest(t, jwtCDN, "POST", uri, checkRequest)
+
+	//Do the request
+	rec := httptest.NewRecorder()
+	router.Mux.ServeHTTP(rec, req)
+	assert.Equal(t, 204, rec.Code)
 }
