+++ sdk/cdsclient/client_workflow.go
@@ -470,22 +470,6 @@ func (c *client) WorkflowRunResultsList(ctx context.Context, projectKey string, 
 	return results, nil
 }
 
-func (c *client) WorkflowRunResultsAdd(ctx context.Context, projectKey string, name string, number int64, addRequest sdk.WorkflowRunResult) error {
-	uri := fmt.Sprintf("/project/%s/workflows/%s/runs/%d/results", projectKey, name, number)
-	if _, err := c.PostJSON(ctx, uri, addRequest, nil); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *client) WorkflowRunResultCheck(ctx context.Context, projectKey string, name string, number int64, apiRef sdk.CDNRunResultAPIRef) error {
-	uri := fmt.Sprintf("/project/%s/workflows/%s/runs/%d/results/check", projectKey, name, number)
-	if _, err := c.PostJSON(ctx, uri, apiRef, nil); err != nil {
-		return err
-	}
-	return nil
-}
-
 func (c *client) WorkflowCachePull(projectKey, integrationName, ref string) (io.Reader, error) {
 	uri := fmt.Sprintf("/project/%s/storage/%s", projectKey, integrationName)
 	store := new(sdk.ArtifactsStore)
